requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-pinecone
langgraph
pinecone-client
pypdf
python-dotenv
streamlit